Refactor block master sync
