+++ core/server/worker/src/main/java/alluxio/worker/block/DefaultBlockWorker.java
@@ -208,7 +208,8 @@ public class DefaultBlockWorker extends AbstractWorker implements BlockWorker {
 
     // Setup BlockMasterSync
     BlockMasterSync blockMasterSync = mResourceCloser
-        .register(new BlockMasterSync(this, mWorkerId, mAddress, mBlockMasterClientPool));
+        .register(new BlockMasterSync(
+            this, mWorkerId, mAddress, mBlockMasterClientPool, mHeartbeatReporter));
     getExecutorService()
         .submit(new HeartbeatThread(HeartbeatContext.WORKER_BLOCK_SYNC, blockMasterSync,
             (int) Configuration.getMs(PropertyKey.WORKER_BLOCK_HEARTBEAT_INTERVAL_MS),
@@ -324,9 +325,12 @@ public class DefaultBlockWorker extends AbstractWorker implements BlockWorker {
     return mBlockStore.createBlockWriter(sessionId, blockId);
   }
 
-  @Override
-  public BlockHeartbeatReport getReport() {
-    return mHeartbeatReporter.generateReport();
+  /**
+   * @return the block heartbeat reporter
+   */
+  @VisibleForTesting
+  public BlockHeartbeatReporter getReporter() {
+    return mHeartbeatReporter;
   }
 
   @Override
